Рефакторинг.
Команды вынесены в классы. Декомозиция. Стилизация.
