+++ commands/Quiz/add.command.php
@@ -1,0 +1,7 @@
+<?php
+
+use SoerBot\Commands\Quiz\Implementations\QuizAddCommand;
+
+return function ($client) {
+    return new QuizAddCommand($client);
+};
