+++ commands/Quiz/quiz.command.php
@@ -1,0 +1,7 @@
+<?php
+
+use SoerBot\Commands\Quiz\Implementations\QuizCommand;
+
+return function ($client) {
+    return new QuizCommand($client);
+};
