+++ commands/System/deploy.command.php
@@ -1,27 +1,7 @@
 <?php
 
-return function ($client) {
-    return new class($client) extends \CharlotteDunois\Livia\Commands\Command {
-        public function __construct(\CharlotteDunois\Livia\LiviaClient $client)
-        {
-            parent::__construct($client, [
-                'name' => 'deploy',
-                'aliases' => [],
-                'group' => 'utils',
-                'description' => 'Deploy command',
-                'guildOnly' => false,
-                'ownerOnly' => true,
-                'throttling' => [
-                    'usages' => 5,
-                    'duration' => 10,
-                ],
-                'guarded' => true,
-            ]);
-        }
+use SoerBot\Commands\System\DeployCommand;
 
-        public function run(\CharlotteDunois\Livia\CommandMessage $message, \ArrayObject $args, bool $fromPattern)
-        {
-            $this->client->emit('stop');
-        }
-    };
+return function ($client) {
+    return new DeployCommand($client);
 };
